# faster rand(::MersenneTwister, UnitRange{<:BitInteger}) (fix #22600)