This is a solution simirar to #14772: generation with `MersenneTwister` is fast, so it's cheaper to waste entropy bits rather than using divisions to save those bits as much as we can (it may have been different when `MersenneTwister` couldn't use the fast SIMD-optimized array functions).

I intend to extend this work (when applicable) before the next release, but this first step will probably be the most visible.

Few timings I get (using in fact `@btime` from BenchmarkTools, and `m` being a `MersenneTwister` object), showing master/PR, or master/`randi`/PR when applicable (using `randi` from StatsBase):

```
julia> @time rand(m, false:true)
21ns / 9ns

julia> @time rand(m, 1:2)
33ns / 21ns / 8.5ns

julia> @time rand(m, 1:2^30)
33ns / 29ns / 8.5ns

julia> @time rand(m, 1:2^30+1)
37ns / 29ns / 19ns
 
julia> @time rand(m, 1:2^52)
38ns / 31ns / 8.5ns

julia> @time rand(m, 1:2^52+1)
38ns / 31ns / 28ns

julia> @time rand(m, 1:2^53)
38ns / 31ns / 13ns

julia> @time rand(m, 1:Int128(2)^104)
102ns / 19ns

julia> @time rand(m, 1:Int128(2)^104+1)
101ns / 39ns

julia> @time rand(m, 1:Int128(2)^105)
102ns / 24ns
```
Basically, with the rejection sampling in this PR, when the lenght of the interval is a bit above a power of 2, then about 1 sample over two is rejected, which explains the change of performances around a power of 2. Also, this takes advantage of the fact that `MersenneTwister` nativily generates 52 bits of entropy, hence a change in performances after lengths of 2^52 and 2^104.