@nalimilan 2017-06-29
Sorry, I'm really not qualified to comment on the RNG code.