@GregPlowman 2017-06-29
Would this function also benefit from using the new method for `UnitRange`, rather than using `RangeGenerator`? 
```
function rand!(rng::AbstractRNG, A::AbstractArray, r::AbstractArray)
     g = RangeGenerator(1:(length(r)))
     for i in eachindex(A)
         @inbounds A[i] = r[rand(rng, g)]
     end
     return A
 end
```
Or perhaps a new RangeGenerator appropriate for new method?