@rfourquet 2017-06-30
> Would this function also benefit from using the new method for UnitRange, rather than using RangeGenerator

According to my preliminary testings, mostly, but with less speed-up and even a slow-down in some cases, e.g. for the range `1:2^52+1` (when calling `rand!(m, A, 1:2^52+1)`). So I'm not yet sure what's the best strategy. Concerning modifying `RangeGenerator`, it would also require checking that this new version is allways faster for all (most) the RNGs types we can test, or to introduce a trait for different RNGs, and call e.g. `RangeGenerator(1:2, rng)`, to choose the best implementation depending on the `rng` type. It will require a bit more time, so I prefered proposing first the simplest fix (which is unambiguously faster).