@mschauer 2017-07-06
Probably we get only one shot at changing this so I think it is worthwhile to think a bit about the problem space.

Basically we have to keep track of three sources of cost:
1. computation of the generator (the cost of ``mm(n,k) = div(n,k)*k``)
2. resampling (e.g. ``rand_ui52_raw``)
3. modulo operation 

in your approach all costs are source 2, previously dominant costs are 1. and 3 and this might well be the right solution.

On this machine, costs for ``mm``, ``rand_ui52_raw`` and ``% k`` are roughly proportional to 5 ; 1 and 1/4.

There is not much what can be done about 3, but (on this machine) it is also only 1/4 of the sampling costs, so worthwhile to do if it decreases the number of rejections by a factor.

``mm(n,k)``  is expensive so it is worthwhile if more than say a handful random numbers are needed. As a basic use case is sampling random integeras from all ranges from 1:1 to 1:n your pull request might be the right thing to do. I just wanted to point out that in this case also the mm(n,k) computation becomes cheaper. The basic observation is that the mm(n,k)  can be computed recursivley. Look at 

I have a "proof of concept" implemenation here https://gist.github.com/mschauer/d6a7a091bada2a828e1ac74e7baa106b

It defines an iterator which counts downwards from ``i = k, k-1, k-2`` and gives  pairs of ``(i, rand(1:i))`` in each iteration. It is cheap if ``k`` is small and i is close to ``k``.

So in praxis one might divide ``k`` by a factor and restart the iterator when ``i`` becomes relatively small. I have not fully understood this yet but wanted to invite you to think about it, in the meanwhile I continue reading here.