@ggggggggg 2017-07-06
If this is PR/master, doesn't this show that master is faster than your PR?
```
julia> @time rand(m, false:true)
21ns / 9ns
```