@rfourquet 2017-07-06
> Probably we get only one shot at changing this

What do you mean? 

> The basic observation is that the mm(n,k) can be computed recursivley.

Very interesting! I will take time to read your proof of concept, and compare efficiency. That said, I guess the simplicity of this PR may be preferable for the general case. 

> If this is PR/master, doesn't this show that master is faster than your PR?

My mistake, thanks for reporting (editing OP).