@mschauer 2017-07-06
> one shot

There are some reasons not to change the random number generator to liberally, because old versions (at least say from julia 1.0 onwards) should be available for reproducibility of numerical experiments. R still has a normal variate generator called "Buggy Kinderman-Ramage" (kept for reproducibility...)