@rfourquet 2017-07-07
> old versions (at least say from julia 1.0 onwards) should be available for reproducibility of numerical experiments

I was vaguely thinking of this reason, but didn't remember having seen this stated clearly in the forums. Do you think the generators could be changed (i.e. the reproducibility broken) between 1.0 and 2.0 ?