@StefanKarpinski 2017-07-11
I would *really* like for RNG to be explicitly set when `srand` is called so that we can change RNGs and other algorithms in the future without breaking code. As a second-best, if we could change them without breaking code silently, that would also be acceptable. The main problem with this is that we don't generally pass RNGs around explicitly – we often use the global default RNG whose type cannot be changed. This seems like something that we very much need a solution to.