@rfourquet 2017-07-11
@StefanKarpinski I'm afraid I don't understand your concern here: in which case code will be broken?