@rfourquet 2017-07-12
I think I understand a bit better (this also relates to point 3. of https://github.com/JuliaLang/julia/issues/17918#issuecomment-269131293), but not fully yet. But by "breaking code", do you mean that the stream of random numbers is modified when we change the implementation (i.e. non-reproducibility) ?  Would you like to prevent the use of `srand([seed])`, and force something like `srand(MersenneTwister, [seed])` to explicitly tell which RNG type to use as the implicit one? then, if the implementation changes, the old one is kept with the old name, and the new updated RNG is named e.g. `MersenneTwister2`. Then, we could still allow `srand()`, with proper documentation, meaning whatever best current  RNG is available as a default, when the user doesn't care about this problem.