@rfourquet 2017-07-12
I feel incline to merge this soon, as that's the best we have for now, while still exploring @mschauer idea.