@tkelman 2017-07-12
are there benchmarks that would track this? @nanosoldier `runbenchmarks(ALL, vs = ":master")`