@rfourquet 2017-07-12
> are there benchmarks that would track this?

under review, see https://github.com/JuliaCI/BaseBenchmarks.jl/pull/82.