@nanosoldier 2017-07-12
[Your benchmark job](https://github.com/JuliaLang/julia/pull/22612#issuecomment-314742438) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/c6087c6c6bee2fed065371c795c6029a712523f8/6924699_vs_0d83de9/report.md). cc @ararslan