@rfourquet 2017-08-11
@nanosoldier `runbenchmarks("random", vs = ":master")`