@nanosoldier 2017-08-11
[Your benchmark job](https://github.com/JuliaLang/julia/pull/22612#issuecomment-321784968) has completed - no performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/30601d128134b9630d7a62800577a0b1a11ce46f/c1e8aaf_vs_55ab3e8/report.md). cc @ararslan